DL-Task/requirements.txt
torch==2.3.0
torchvision==0.18.0
Pillow
pandas
matplotlib
opencv-python
torchsummary
tqdm
